Reorganiza tests, corrije dependencia circular, mueve inicializacion de usuarios.
